requirements.txt
snowflake-snowpark-python
snowflake-connector-python
streamlit
ragas
datasets
langchain-anthropic
langchain-huggingface
sentence-transformers
anthropic